[AIRFLOW-6368] Move conf tests from test_core to test_configuration
